(main) [6] [Andy Malik]
Once I was able to test without uploading to github, I uh, did the thing. Followed a video to help build the 3D world and tried to mess with my own animations and interactions. Click the ball, and hold down the mouse (you may drag off the ball as long as the mouse is down) and the ball bounces. Still a few flaws but good work for one night.